[godot] Addresses #2899, #2980, #2985
Addresses the loader/saver registration issue in #2899.

Additionally, makes it such that you can adjust the SpineSkeletonData in
the inspector without crashes. This appears to have been caused by a
dangling pointer.

Finally, double-clicking on JSON in the inspector opens the text editor
and does not crash the Godot editor.